requirements.txt
# TORCH ON WINDOWS REQUIRES PYTHON 3.11!

# System
python-dotenv
whitenoise
soundfile

# AI components
gpt4all
SpeechRecognition
PyAudio
git+https://github.com/suno-ai/bark.git

# Django
Django
asgiref
autopep8
dj-database-url
gunicorn
pycodestyle
pytz
sqlparse
toml
django-environ
django-pipeline
psycopg2
css-html-js-minify

# Theme
django-theme-material-kit==1.0.19
